requirements.txt
flask
Flask-BasicAuth
gunicorn
pydantic
python-telegram-bot